+++ app/src/main/java/com/treasurehunt/ui/profile/adapter/FriendAdapter.kt
@@ -1,4 +1,4 @@
-package com.treasurehunt.ui.profile
+package com.treasurehunt.ui.profile.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -10,6 +10,7 @@ import com.bumptech.glide.Glide
 import com.treasurehunt.R
 import com.treasurehunt.databinding.ItemFriendBinding
 import com.treasurehunt.ui.model.UserModel
+import com.treasurehunt.ui.profile.FriendClickListener
 
 class FriendAdapter(
     private val friendClickListener: FriendClickListener,
